Shorten clade names in a dataset to chosen taxonomic level

This function shortens the taxonomic names of clades in a given dataset based on a specified taxonomic level, using '|' as a separator.

Usage

shorten_clade_names(
  data,
  taxa_lvl,
  apply_to_colnames = TRUE,
  selected_cols = NULL
)

Arguments

data

The input dataset. Assumes clade names are of the structure k__Bacteria|p__Firmicutes|c__Bacilli|o__Lactobacillales|f__Lactobacillaceae and found in the 'clade_name' column.

taxa_lvl

The taxonomic level at which the clade names should be shortened. Valid options include 'kingdom', 'phylum', 'class', 'order', 'family', 'genus', or 'species'. First letter abbreviations (e.g., 's') are also accepted.

apply_to_colnames

Logical indicating whether the shortening should be applied to column names or row values. Default is TRUE.

selected_cols

A character vector specifying the columns to which the shortening should be applied. If NULL, the shortening is applied to all columns. Default is NULL.

Value

The dataset with the clade names shortened based on the specified taxonomic level.

Details

This function uses the dplyr package for data manipulation.

Author

Jérémy Rotzetter

Examples

head(species_only$clade_name)
#> [1] "k__Bacteria|p__Proteobacteria|c__Gammaproteobacteria|o__Moraxellales|f__Moraxellaceae|g__Moraxella|s__Moraxella_nonliquefaciens"                     
#> [2] "k__Bacteria|p__Actinobacteria|c__Actinomycetia|o__Corynebacteriales|f__Corynebacteriaceae|g__Corynebacterium|s__Corynebacterium_pseudodiphtheriticum"
#> [3] "k__Bacteria|p__Firmicutes|c__Negativicutes|o__Veillonellales|f__Veillonellaceae|g__Veillonella|s__Veillonella_dispar"                                
#> [4] "k__Bacteria|p__Bacteroidota|c__Bacteroidia|o__Bacteroidales|f__Prevotellaceae|g__Prevotella|s__Prevotella_histicola"                                 
#> [5] "k__Bacteria|p__Firmicutes|c__Bacilli|o__Lactobacillales|f__Streptococcaceae|g__Streptococcus|s__Streptococcus_mitis"                                 
#> [6] "k__Bacteria|p__Proteobacteria|c__Gammaproteobacteria|o__Pasteurellales|f__Pasteurellaceae|g__Haemophilus|s__Haemophilus_haemolyticus"                

taxa_shortened <- shorten_clade_names(
  species_only,
  "species",
  apply_to_colnames = FALSE,
  selected_cols = "clade_name")

  head(taxa_shortened$clade_name)
#> [1] "Moraxella_nonliquefaciens"           
#> [2] "Corynebacterium_pseudodiphtheriticum"
#> [3] "Veillonella_dispar"                  
#> [4] "Prevotella_histicola"                
#> [5] "Streptococcus_mitis"                 
#> [6] "Haemophilus_haemolyticus"